modify Entity

+++ Server/src/main/java/com/threeracha/gaewoonhae/db/domain/Emoji.java
@@ -17,7 +17,7 @@ public class Emoji {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="emoji_id")
-    private int emojiId;
+    private Long emojiId;
 
     @Column(name="emoji_price", nullable = false)
     private int emojiPrice;
